waypoint ckin

+++ include/Camera.h
@@ -20,7 +20,7 @@ namespace vt {
 class FrameBuffer;
 
 class Camera : public TransformObject,
-               public FrameObject<glm::vec2, float>
+               public FrameObject<glm::ivec2, int>
 {
 public:
     enum projection_mode_t {
@@ -32,9 +32,9 @@ public:
            glm::vec3         origin          = glm::vec3(0),
            glm::vec3         target          = glm::vec3(-1),
            float             fov             = DEFAULT_FOV,
-           glm::vec2         offset          = glm::vec2(0),
-           glm::vec2         dim             = glm::vec2(DEFAULT_VIEWPORT_WIDTH,
-                                                         DEFAULT_VIEWPORT_HEIGHT),
+           glm::ivec2        offset          = glm::ivec2(0),
+           glm::ivec2        dim             = glm::ivec2(DEFAULT_VIEWPORT_WIDTH,
+                                                          DEFAULT_VIEWPORT_HEIGHT),
            float             near_plane      = DEFAULT_NEAR_PLANE,
            float             far_plane       = DEFAULT_FAR_PLANE,
            glm::vec2         ortho_dim       = glm::vec2(DEFAULT_ORTHO_VIEWPORT_WIDTH,
@@ -63,7 +63,7 @@ public:
     }
     void set_fov(float fov);
 
-    void resize(float left, float bottom, float width, float height);
+    void resize(int left, int bottom, int width, int height);
 
     float get_near_plane() const
     {
