+++ src/Camera.cpp
@@ -19,8 +19,8 @@ Camera::Camera(std::string       name,
                glm::vec3         origin,
                glm::vec3         target,
                float             fov,
-               glm::vec2         offset,
-               glm::vec2         dim,
+               glm::ivec2        offset,
+               glm::ivec2        dim,
                float             near_plane,
                float             far_plane,
                glm::vec2         ortho_dim,
@@ -108,9 +108,9 @@ void Camera::set_fov(float fov)
     mark_dirty_transform();
 }
 
-void Camera::resize(float left, float bottom, float width, float height)
+void Camera::resize(int left, int bottom, int width, int height)
 {
-    FrameObject<glm::vec2, float>::resize(left, bottom, width, height);
+    FrameObject<glm::ivec2, int>::resize(left, bottom, width, height);
     m_is_dirty_projection_transform = true;
     mark_dirty_transform();
 }
